ClientDepthCam/requirements.txt
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
Flask==2.3.3
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pyrealsense2==2.54.1.5217
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.0
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.7
